+++ src/vs/platform/ipc/electron-sandbox/services.ts
