+++ src/vs/workbench/services/diagnostics/electron-sandbox/diagnosticsService.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { registerSharedProcessRemoteService } from 'vs/platform/ipc/common/services';
+import { registerSharedProcessRemoteService } from 'vs/platform/ipc/electron-sandbox/services';
 import { IDiagnosticsService } from 'vs/platform/diagnostics/common/diagnostics';
 
 registerSharedProcessRemoteService(IDiagnosticsService, 'diagnostics', { supportsDelayedInstantiation: true });
