+++ src/vs/workbench/services/extensionManagement/electron-sandbox/extensionTipsService.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
-import { ISharedProcessService } from 'vs/platform/ipc/common/services';
+import { ISharedProcessService } from 'vs/platform/ipc/electron-sandbox/services';
 import { IChannel } from 'vs/base/parts/ipc/common/ipc';
 import { IExtensionTipsService, IExecutableBasedExtensionTip, IWorkspaceTips, IConfigBasedExtensionTip } from 'vs/platform/extensionManagement/common/extensionManagement';
 import { URI } from 'vs/base/common/uri';
